# specs/001-empleados-clave-compuesta/contracts/empleados.openapi.yaml
openapi: 3.0.3
info:
  title: API Empleados - Clave Compuesta
  version: 1.0.0
  description: Contrato para gestión de empleados con clave `EMP-<consecutivo>`.
servers:
  - url: http://localhost:8080
security:
  - basicAuth: []

paths:
  /api/v1/empleados:
    post:
      summary: Crear empleado
      operationId: crearEmpleado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoCreateRequest'
      responses:
        '201':
          description: Empleado creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          description: Datos inválidos o intento de clave manual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Colisión de clave por concurrencia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
    get:
      summary: Listar empleados
      operationId: listarEmpleados
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              pattern: '^(clave|nombre|direccion|telefono),(asc|desc)$'
          style: form
          explode: true
      responses:
        '200':
          description: Lista paginada de empleados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEmpleadoResponse'
        '401':
          description: No autenticado
        '422':
          description: Parámetros de paginación/ordenamiento inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/empleados/{clave}:
    get:
      summary: Consultar empleado por clave
      operationId: consultarEmpleadoPorClave
      parameters:
        - name: clave
          in: path
          required: true
          schema:
            type: string
            pattern: '^EMP-[0-9]+$'
      responses:
        '200':
          description: Empleado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '404':
          description: Empleado no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
    put:
      summary: Actualizar empleado
      operationId: actualizarEmpleado
      parameters:
        - name: clave
          in: path
          required: true
          schema:
            type: string
            pattern: '^EMP-[0-9]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoUpdateRequest'
      responses:
        '200':
          description: Empleado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Empleado no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
    delete:
      summary: Eliminar empleado
      operationId: eliminarEmpleado
      parameters:
        - name: clave
          in: path
          required: true
          schema:
            type: string
            pattern: '^EMP-[0-9]+$'
      responses:
        '204':
          description: Empleado eliminado
        '404':
          description: Empleado no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  schemas:
    EmpleadoCreateRequest:
      type: object
      required: [nombre, direccion, telefono]
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        direccion:
          type: string
          minLength: 1
          maxLength: 100
        telefono:
          type: string
          minLength: 1
          maxLength: 100
        clave:
          type: string
          description: No permitido en create

    EmpleadoUpdateRequest:
      type: object
      required: [nombre, direccion, telefono]
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        direccion:
          type: string
          minLength: 1
          maxLength: 100
        telefono:
          type: string
          minLength: 1
          maxLength: 100

    EmpleadoResponse:
      type: object
      required: [clave, nombre, direccion, telefono]
      properties:
        clave:
          type: string
          pattern: '^EMP-[0-9]+$'
          example: EMP-1001
        nombre:
          type: string
        direccion:
          type: string
        telefono:
          type: string

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string

    PaginatedEmpleadoResponse:
      type: object
      required: [content, number, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EmpleadoResponse'
        number:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0